// packages/component-library/src/core/DataPointsTable/schemas/DeviceConfigurationSchema.ts
// @ts-nocheck
import { z } from 'zod';
import { DeviceProtocolSchema } from './ProtocolSchemas';
import EdgeXAPI from '../api/EdgeXAPI';

// This schema defines the structure and validation rules for a Device Configuration object,
// as supported by EdgeX. It includes fields for device metadata, state, associated service and profile names,
// communication protocols, and optional auto-events configuration.
// The schema ensures that required fields are present and conform to expected formats and constraints.
export const DeviceConfigurationSchema = z.object({
    // The device name is required and needs to be unique across EdgeX.
    // TODO: Determine if we want an async refinement here to query EdgeX for existence.
    name: z
        .string('Device name is required')
        .min(1, 'Device name cannot be empty')
        .max(66, 'Device name must be at most 66 characters')
        .regex(/^[a-zA-Z0-9-_]+$/, 'Device name can only contain alphanumeric characters, hyphens, and underscores'),
    // The id, created, modified fields are generated by EdgeX and are thus optional.
    id: z.string().optional(),
    created: z.number().optional(),
    modified: z.number().optional(),
    // The description and labels are optional metadata fields.
    description: z.string().optional(),
    labels: z.array(z.string()).optional(),
    // The adminState and operatingState are managed by EdgeX, with default values provided.
    adminState: z.enum(['LOCKED', 'UNLOCKED']).default('UNLOCKED'),
    operatingState: z.enum(['UP', 'DOWN']).default('UP'),
    // The device service name is optional, as we assign a default service if not provided.
    serviceName: z.string().nonempty('Name cannot be empty.').default('smp-device-service'),
    // The profile name is required to link the device to its profile.
    profileName: z
        .string('Device Profile name is required')
        .min(1, 'Device Profile name is required')
        .refine(async (name) => {
            // Check if the profile exists in the system
            try {
                const profile = await EdgeXAPI.getDeviceProfile(name);
                return !!profile;
                // eslint-disable-next-line @typescript-eslint/no-unused-vars
            } catch (e) {
                return false;
            }
        }, 'Device Profile does not exist'),
    // The protocols field is required and must conform to the DeviceProtocolSchema.
    protocols: DeviceProtocolSchema,
    // TODO: Implement auto-events schema if that ever becomes relevant.
    autoEvents: z.array(z.any()).optional(),
});
export type DeviceConfiguration = z.infer<typeof DeviceConfigurationSchema>;